pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "decision-trace"
version = "0.1.0"
description = "Decision Trace v1 SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pydantic>=2.0", "jsonschema>=4.0", "PyYAML>=6.0", "opentelemetry-api>=1.20", "opentelemetry-sdk>=1.20"]

[project.optional-dependencies]
test = ["pytest>=7.0", "fastapi>=0.110", "httpx>=0.24"]
collector = ["fastapi>=0.110", "uvicorn>=0.23"]

[project.scripts]
decision-trace = "decision_trace.cli:main"
decision-trace-collector = "decision_trace_collector.main:main"

[tool.setuptools.packages.find]
where = ["sdk/python", "services/collector"]
include = ["decision_trace*", "decision_trace_collector*"]

[tool.pytest.ini_options]
testpaths = ["sdk/python/tests", "services/collector/tests"]
